refactored pipeline to use __main__ for command line execution and PEP8 file naming
